Rename al archivo config_db_users donde se hace referencia a la db de usuarios.
